execution/stagedsync: prepend held-back retries via slices.Insert, keep ensureLen, assert precondition
